+++ device-plugin/pkg/mlu/const.go
@@ -19,17 +19,18 @@ import pluginapi "k8s.io/kubelet/pkg/apis/deviceplugin/v1beta1"
 type pluginMode string
 
 const (
+	cambriconVisibleDevices = "CAMBRICON_VISIBLE_DEVICES"
+
 	DsmluLockTime           = "cambricon.com/dsmlu.lock"
 	DsmluProfile            = "CAMBRICON_DSMLU_PROFILE"
 	DsmluProfileAndInstance = "CAMBRICON_DSMLU_PROFILE_INSTANCE"
-	DsmluResourceAssigned   = "CAMBRICON_DSMLU_ASSIGHED"
+	DsmluResourceAssigned   = "CAMBRICON_DSMLU_ASSIGNED"
 
-	mluLinkPolicyUnsatisfied = "mluLinkPolicyUnsatisfied"
-
-	normalMlu  = "mlu"
-	realCounts = "real-mlu-counts"
-	retries    = 5
-	serverSock = pluginapi.DevicePluginPath + "cambricon.sock"
+	normalMlu      = "mlu"
+	realCounts     = "real-mlu-counts"
+	retries        = 5
+	serverSock     = pluginapi.DevicePluginPath + "cambricon.sock"
+	virtualDevices = "VIRTUAL_DEVICES"
 )
 
 const (