+++ device-plugin/pkg/mlu/server.go
@@ -129,7 +129,7 @@ func (m *CambriconDevicePlugin) Start() error {
 	}
 	conn.Close()
 
-	if !m.options.DisableHealthCheck {
+	if !m.options.DisableHealthCheck && m.profile != realCounts {
 		go m.healthcheck()
 	}
 
@@ -175,7 +175,7 @@ func (m *CambriconDevicePlugin) Register(kubeletEndpoint, resourceName string) e
 }
 
 // ListAndWatch lists devices and update that list according to the health status
-func (m *CambriconDevicePlugin) ListAndWatch(e *pluginapi.Empty, s pluginapi.DevicePlugin_ListAndWatchServer) error {
+func (m *CambriconDevicePlugin) ListAndWatch(_ *pluginapi.Empty, s pluginapi.DevicePlugin_ListAndWatchServer) error {
 	if err := s.Send(&pluginapi.ListAndWatchResponse{Devices: m.devs}); err != nil {
 		log.Errorf("Failed send list and watch response via sock %s with %v", m.socket, err)
 		syscall.Kill(os.Getpid(), syscall.SIGHUP)
@@ -202,9 +202,41 @@ func (m *CambriconDevicePlugin) ListAndWatch(e *pluginapi.Empty, s pluginapi.Dev
 	}
 }
 
-func (m *CambriconDevicePlugin) PrepareResponse(uuids []string) pluginapi.ContainerAllocateResponse {
+func (m *CambriconDevicePlugin) PrepareResponse(uuids []string) *pluginapi.ContainerAllocateResponse {
+	resp := &pluginapi.ContainerAllocateResponse{}
 
-	resp := pluginapi.ContainerAllocateResponse{}
+	devPaths := m.uuidToPath(uuids)
+	log.Debugf("Prepare response device paths %v", devPaths)
+
+	if m.options.EnabledCDI {
+		if (m.options.Mode == Mim || m.options.Mode == DynamicSmlu) && m.profile != normalMlu {
+			// TODO: support CDI in mim or dynamic-smlu mode
+			log.Error("Not support CDI in mim or dynamic-smlu mode")
+			return resp
+		}
+		cdiDevices := []*pluginapi.CDIDevice{}
+		for _, uuid := range uuids {
+			deviceID := "mlu-" + strconv.Itoa(int(m.devsInfo[uuid].Slot))
+			cdiName := fmt.Sprintf("cambricon.com/mlu=%s", deviceID)
+			cdiDevices = append(cdiDevices, &pluginapi.CDIDevice{Name: cdiName})
+		}
+		resp.CDIDevices = cdiDevices
+		return resp
+	}
+
+	if m.options.UseRuntime {
+		resp.Envs = make(map[string]string)
+		if (m.options.Mode == Mim || m.options.Mode == DynamicSmlu) && m.profile != normalMlu {
+			resp.Envs[virtualDevices] = strings.Join(devPaths, ";")
+			return resp
+		}
+		var slots []string
+		for _, uuid := range uuids {
+			slots = append(slots, strconv.Itoa(int(m.devsInfo[uuid].Slot)))
+		}
+		resp.Envs[cambriconVisibleDevices] = strings.Join(slots, ",")
+		return resp
+	}
 
 	if m.options.MountRPMsg {
 		resp.Mounts = append(resp.Mounts, &pluginapi.Mount{
@@ -221,62 +253,59 @@ func (m *CambriconDevicePlugin) PrepareResponse(uuids []string) pluginapi.Contai
 		})
 	}
 
-	devpaths := m.uuidToPath(uuids)
-	log.Debugf("prepareresponse devicepaths %v", devpaths)
-
 	if m.deviceList.hasCtrlDev {
-		addDevice(&resp, mluMonitorDeviceName)
+		addDevice(resp, mluMonitorDeviceName)
 	}
 
 	if m.deviceList.hasGdrDev {
-		addDevice(&resp, mluGdrDeviceName)
+		addDevice(resp, mluGdrDeviceName)
 	}
 
 	omitDup := map[string]struct{}{}
-	for _, devpath := range devpaths {
+	for _, devPath := range devPaths {
 		if (m.options.Mode == Mim || m.options.Mode == DynamicSmlu) && m.profile != normalMlu {
 			// devpath is like "/dev/cambricon_dev0,/dev/cambricon_ipcm0,/dev/cambricon-caps/cap_dev0_mi1"
-			pathSets := strings.Split(devpath, ",")
+			pathSets := strings.Split(devPath, ",")
 			if len(pathSets) != 3 {
-				log.Printf("invalid devpath %s", devpath)
+				log.Printf("Invalid devPath %s", devPath)
 				continue
 			}
 			if _, ok := omitDup[pathSets[0]]; !ok {
-				addDevice(&resp, pathSets[0])
+				addDevice(resp, pathSets[0])
 				if m.deviceList.hasIpcmDev {
-					addDevice(&resp, pathSets[1])
+					addDevice(resp, pathSets[1])
 				}
 				omitDup[pathSets[0]] = struct{}{}
 			}
-			addDevice(&resp, pathSets[2])
+			addDevice(resp, pathSets[2])
 			continue
 		}
 
 		var index int
-		_, err := fmt.Sscanf(devpath, mluDeviceName+"%d", &index)
+		_, err := fmt.Sscanf(devPath, mluDeviceName+"%d", &index)
 		if err != nil {
 			log.Printf("Failed to get device index for device path %v", err)
 			continue
 		}
 		if m.deviceList.hasMsgqDev {
-			addDevice(&resp, fmt.Sprintf(mluMsgqDeviceName+":%d", index))
+			addDevice(resp, fmt.Sprintf(mluMsgqDeviceName+":%d", index))
 		}
 		if m.deviceList.hasRPCDev {
-			addDevice(&resp, fmt.Sprintf(mluRPCDeviceName+":%d", index))
+			addDevice(resp, fmt.Sprintf(mluRPCDeviceName+":%d", index))
 		}
 		if m.deviceList.hasCmsgDev {
-			addDevice(&resp, fmt.Sprintf(mluCmsgDeviceName+"%d", index))
+			addDevice(resp, fmt.Sprintf(mluCmsgDeviceName+"%d", index))
 		}
 		if m.deviceList.hasCommuDev {
-			addDevice(&resp, fmt.Sprintf(mluCommuDeviceName+"%d", index))
+			addDevice(resp, fmt.Sprintf(mluCommuDeviceName+"%d", index))
 		}
 		if m.deviceList.hasIpcmDev {
-			addDevice(&resp, fmt.Sprintf(mluIpcmDeviceName+"%d", index))
+			addDevice(resp, fmt.Sprintf(mluIpcmDeviceName+"%d", index))
 		}
 		if m.deviceList.hasUARTConsoleDev && m.options.EnableConsole {
-			addDevice(&resp, fmt.Sprintf(mluUARTConsoleDeviceName+"%d", index))
+			addDevice(resp, fmt.Sprintf(mluUARTConsoleDeviceName+"%d", index))
 		}
-		addDevice(&resp, devpath)
+		addDevice(resp, devPath)
 	}
 	return resp
 }
@@ -319,7 +348,7 @@ func (m *CambriconDevicePlugin) allocateDynamicSmlu(ctx context.Context) (*plugi
 		}
 		_, err = m.clientset.CoreV1().Pods(pod.Namespace).Patch(ctx, pod.Name, types.StrategicMergePatchType, patchedAnnotation, metav1.PatchOptions{})
 		for i := 0; i < retries && err != nil; i++ {
-			log.Warnf("patchPodAnnotation err: %v, retried times: %d", err, i)
+			log.Warnf("PatchPodAnnotation err: %v, retried times: %d", err, i)
 			time.Sleep(time.Duration(rand.Intn(i)) * 10 * time.Millisecond)
 			_, err = m.clientset.CoreV1().Pods(pod.Namespace).Patch(ctx, pod.Name, types.StrategicMergePatchType, patchedAnnotation, metav1.PatchOptions{})
 		}
@@ -333,7 +362,7 @@ func (m *CambriconDevicePlugin) allocateDynamicSmlu(ctx context.Context) (*plugi
 			err = m.releaseNodeLock()
 		}
 		if err != nil {
-			log.Printf("releaseNodeLock exceeds retry count %d", retries)
+			log.Printf("ReleaseNodeLock exceeds retry count %d", retries)
 		}
 		return resp, nil
 	}
@@ -411,18 +440,18 @@ func (m *CambriconDevicePlugin) allocateDynamicSmlu(ctx context.Context) (*plugi
 		Profile: strings.ReplaceAll(dsmluInfo.Name, "+", "-"),
 	}
 
-	responses := pluginapi.AllocateResponse{}
+	responses := &pluginapi.AllocateResponse{}
 	resp := m.PrepareResponse([]string{uid})
-	responses.ContainerResponses = append(responses.ContainerResponses, &resp)
+	responses.ContainerResponses = append(responses.ContainerResponses, resp)
 
 	dynamicSmlu = make(map[string]*pluginapi.AllocateResponse)
-	dynamicSmlu[pod.Name] = &responses
+	dynamicSmlu[pod.Name] = responses
 	profileAndInstance = make(map[string]string)
 	profileAndInstance[pod.Name] = fmt.Sprintf("%d_%d_%d_%d", profileID, mluIntance, pl.Slot, dsmluInfo.InstanceID)
 
 	log.Debugf("Store in dynamicSmlu Map with key %s, value %v", pod.Name, responses)
 
-	return &responses, nil
+	return responses, nil
 }
 
 // Allocate which return list of devices.
@@ -442,7 +471,7 @@ func (m *CambriconDevicePlugin) Allocate(ctx context.Context, reqs *pluginapi.Al
 			}
 		}
 		car := m.PrepareResponse(req.DevicesIDs)
-		responses.ContainerResponses = append(responses.ContainerResponses, &car)
+		responses.ContainerResponses = append(responses.ContainerResponses, car)
 	}
 
 	tf := time.Now()
@@ -490,22 +519,16 @@ func (m *CambriconDevicePlugin) healthcheck() {
 // Serve starts the gRPC server and register the device plugin to Kubelet
 func (m *CambriconDevicePlugin) Serve() error {
 	if m.options.CnmonPath != "" && !path.IsAbs(m.options.CnmonPath) {
-		log.Panicf("invalid cnmon path: %s", m.options.CnmonPath)
+		log.Panicf("Invalid cnmon path: %s", m.options.CnmonPath)
 	}
 
 	if m.options.Mode == TopologyAware {
 		m.allocator = allocator.New(m.options.MLULinkPolicy, m.devsInfo)
-		m.clientset = initClientSet()
-
-		if m.options.MLULinkPolicy != bestEffort {
-			if err := m.updateNodeMLULinkAnnotation(0); err != nil {
-				return err
-			}
-		}
+		m.clientset = InitClientSet()
 	}
 
 	if m.options.Mode == DynamicSmlu {
-		m.clientset = initClientSet()
+		m.clientset = InitClientSet()
 		if err := m.releaseNodeLock(); err != nil {
 			return err
 		}
@@ -553,7 +576,7 @@ func (m *CambriconDevicePlugin) Serve() error {
 	return nil
 }
 
-func (m *CambriconDevicePlugin) GetPreferredAllocation(ctx context.Context, r *pluginapi.PreferredAllocationRequest) (*pluginapi.PreferredAllocationResponse, error) {
+func (m *CambriconDevicePlugin) GetPreferredAllocation(_ context.Context, r *pluginapi.PreferredAllocationRequest) (*pluginapi.PreferredAllocationResponse, error) {
 	response := &pluginapi.PreferredAllocationResponse{}
 	for _, req := range r.ContainerRequests {
 		var allocated []string
@@ -562,7 +585,7 @@ func (m *CambriconDevicePlugin) GetPreferredAllocation(ctx context.Context, r *p
 		case EnvShare:
 			allocated, err = getPreferredEnvShareDeviceID(req)
 			if err != nil {
-				log.Errorf("failed to get preferred allocated devices for env share mode, requests %d, err: %v", req.AllocationSize, err)
+				log.Errorf("Failed to get preferred allocated devices for env share mode, requests %d, err: %v", req.AllocationSize, err)
 				return response, err
 			}
 		case TopologyAware:
@@ -570,11 +593,11 @@ func (m *CambriconDevicePlugin) GetPreferredAllocation(ctx context.Context, r *p
 			required := m.getSlots(req.MustIncludeDeviceIDs)
 			allocated, err = m.getPreferredAllocatedDeviceUUIDs(available, required, int(req.AllocationSize))
 			if err != nil {
-				log.Errorf("failed to get preferred allocated devices, available: %v, size: %d, err: %v", available, req.AllocationSize, err)
+				log.Errorf("Failed to get preferred allocated devices, available: %v, size: %d, err: %v", available, req.AllocationSize, err)
 				return response, err
 			}
 		default:
-			log.Errorf("not supported mode %s", m.options.Mode)
+			log.Errorf("Not supported mode %s", m.options.Mode)
 			return response, fmt.Errorf("not supported mode %s", m.options.Mode)
 		}
 		resp := &pluginapi.ContainerPreferredAllocationResponse{
@@ -588,11 +611,11 @@ func (m *CambriconDevicePlugin) GetPreferredAllocation(ctx context.Context, r *p
 func (m *CambriconDevicePlugin) getPreferredAllocatedDeviceUUIDs(available []uint, required []uint, size int) ([]string, error) {
 	// todo: consider required list for init containers and numa. ignore it for now.
 	if len(required) != 0 {
-		log.Printf("required device slice not empty, ignore it. %v", required)
+		log.Printf("Required device slice not empty, ignore it. %v", required)
 	}
 
 	log.Println("=== Start getPreferredAllocatedDeviceUUIDs ===")
-	log.Printf("available devs: %v, size %d", available, size)
+	log.Printf("Available devs: %v, size %d", available, size)
 
 	var devs []uint
 	if m.options.MLULinkPolicy == bestEffort && len(available) == size {
@@ -601,14 +624,11 @@ func (m *CambriconDevicePlugin) getPreferredAllocatedDeviceUUIDs(available []uin
 		var err error
 		devs, err = m.allocator.Allocate(available, required, size)
 		if err != nil {
-			if e := m.updateNodeMLULinkAnnotation(size); e != nil {
-				log.Errorf("updateNodeMLULinkAnnotation err: %v", e)
-			}
 			return nil, err
 		}
 	}
 
-	log.Printf("preferred devices %v", devs)
+	log.Printf("Preferred devices %v", devs)
 
 	uuids := []string{}
 	for _, dev := range devs {
@@ -621,37 +641,6 @@ func (m *CambriconDevicePlugin) getPreferredAllocatedDeviceUUIDs(available []uin
 
 	log.Println("=== Finish getPreferredAllocatedDeviceUUIDs ===")
 	return uuids, nil
-}
-
-func (m *CambriconDevicePlugin) createAnnotationWithTimestamp(size int) error {
-	node, err := m.clientset.CoreV1().Nodes().Get(context.TODO(), m.nodeHostname, metav1.GetOptions{})
-	if err != nil {
-		return fmt.Errorf("get node err %v", err)
-	}
-	if size == 0 {
-		delete(node.Annotations, mluLinkPolicyUnsatisfied)
-	} else {
-		timeStamp := strconv.FormatInt(time.Now().Unix(), 10)
-		if len(node.Annotations) == 0 {
-			node.Annotations = make(map[string]string)
-		}
-		node.Annotations[mluLinkPolicyUnsatisfied] = fmt.Sprintf("%d-%s-%s", size, m.options.MLULinkPolicy, timeStamp)
-	}
-	_, err = m.clientset.CoreV1().Nodes().Update(context.TODO(), node, metav1.UpdateOptions{})
-	if err != nil {
-		return fmt.Errorf("update node err: %v", err)
-	}
-	return nil
-}
-
-func (m *CambriconDevicePlugin) updateNodeMLULinkAnnotation(size int) error {
-	err := m.createAnnotationWithTimestamp(size)
-	for i := 0; i < retries && err != nil; i++ {
-		log.Printf("createAnnotationWithTimestamp err: %v, retried times: %d", err, i+1)
-		time.Sleep(time.Duration(rand.Intn(i)) * 10 * time.Millisecond)
-		err = m.createAnnotationWithTimestamp(size)
-	}
-	return err
 }
 
 func (m *CambriconDevicePlugin) getSlots(ids []string) []uint {
@@ -676,14 +665,14 @@ func addDevice(car *pluginapi.ContainerAllocateResponse, devPath string) {
 	car.Devices = append(car.Devices, dev)
 }
 
-func initClientSet() kubernetes.Interface {
+func InitClientSet() kubernetes.Interface {
 	config, err := rest.InClusterConfig()
 	if err != nil {
-		log.Errorf("Failed to get in cluser config, err: %v", err)
+		log.Panicf("Failed to get in cluster config, err: %v", err)
 	}
 	clientset, err := kubernetes.NewForConfig(config)
 	if err != nil {
-		log.Errorf("Failed to init clientset, err: %v", err)
+		log.Panicf("Failed to init clientset, err: %v", err)
 	}
 	return clientset
 }
@@ -695,7 +684,7 @@ func getPreferredEnvShareDeviceID(req *pluginapi.ContainerPreferredAllocationReq
 		uuid := ids[0]
 		vf, err := strconv.Atoi(ids[len(ids)-1])
 		if err != nil {
-			log.Errorf("convert value to int, error %v, value %s", err, ids[len(ids)-1])
+			log.Errorf("Convert value to int, error %v, value %s", err, ids[len(ids)-1])
 			return nil, err
 		}
 		dMap[uuid] = append(dMap[uuid], vf)